requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.26
requests==2.25.1
python-dotenv==0.19.0

